+++ scriptconfig/value.py
@@ -13,16 +13,32 @@ class Value(ub.NiceRepr):
     the type that the value should be (Used when parsing sys.argv).
 
     Attributes:
-        value (object): A float, int, etc...
-        type (Type): the "type" of the value. This is usually used if the
-            value specified is not the type that `self.value` would usually
-            be set to.
-        parsekw (dict): kwargs for to argparse add_argument
-        position (None | int): if an integer, then we allow this value
-            to be a positional argument in the argparse CLI. Note, that
-            values with the same position index will cause conflicts.
+        value (object):
+            A float, int, etc...
+
+        type (Type):
+            the "type" of the value. This is usually used if the value
+            specified is not the type that `self.value` would usually be set
+            to.
+
+        parsekw (dict):
+            kwargs for to argparse add_argument
+
+        position (None | int):
+            if an integer, then we allow this value to be a positional argument
+            in the argparse CLI. Note, that values with the same position index
+            will cause conflicts. Also note: positions indexes should start
+            from 1.
 
         isflag (bool, default=False): if True, args will be parsed as booleans
+
+        alias (List[str]):
+            other long names (that will be prefixed with '--') that will be
+            accepted by the argparse CLI.
+
+        short_alias (List[str]):
+            other short names (that will be prefixed with '-') that will be
+            accepted by the argparse CLI.
 
     Example:
         >>> self = Value(None, type=float)
@@ -37,7 +53,8 @@ class Value(ub.NiceRepr):
     __scfg_class__ = 'Value'
 
     def __init__(self, value=None, type=None, help=None, choices=None,
-                 position=None, isflag=False, nargs=None, alias=None):
+                 position=None, isflag=False, nargs=None, alias=None,
+                 required=False, short_alias=None):
         self.value = None
         self.type = type
         self.alias = alias
@@ -49,6 +66,8 @@ class Value(ub.NiceRepr):
             'choices': choices,
             'nargs': nargs,
         }
+        self.required = required
+        self.short_alias = short_alias
         self.update(value)
 
     def __nice__(self):